+++ CHANGELOG.md
@@ -10,7 +10,7 @@ All notable changes to this project will be documented in this file.
 ### Fixed
 
 - The reworked `criticalup run` behavior was not correctly checking that the toolchain specified
-  in `criticalup.toml` was installed. This lead to some situations where users could accidently 
+  in `criticalup.toml` was installed. This lead to some situations where users could accidentally 
   run a non-Ferrocene tool when they meant to run Ferrocene tools. This has been fixed.
 
 ### Added