+++ CHANGELOG.md
@@ -11,7 +11,8 @@ All notable changes to this project will be documented in this file.
 
 - The reworked `criticalup run` behavior was not correctly checking that the toolchain specified
   in `criticalup.toml` was installed. This lead to some situations where users could accidentally 
-  run a non-Ferrocene tool when they meant to run Ferrocene tools. This has been fixed.
+  run a non-Ferrocene tool when they meant to run Ferrocene tools. This behavior has been altered
+  and CriticalUp will now present users with an error suggesting they install the toolchain.
 
 ### Added