+++ CHANGELOG.md
@@ -7,6 +7,13 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Fixed
+
+- The reworked `criticalup run` behavior was not correctly checking that the toolchain specified
+  in `criticalup.toml` was installed. This lead to some situations where users could accidentally 
+  run a non-Ferrocene tool when they meant to run Ferrocene tools. This behavior has been altered
+  and CriticalUp will now present users with an error suggesting they install the toolchain.
+
 ### Added
 
 - New subcommand `init` creates a simple project manifest file in the current directory. A flag `--print` can